+++ docs/versionhistory.rst
@@ -13,6 +13,7 @@ APScheduler, see the :doc:`migration section <migration>`.
 - Changed ``from_async_sqla_engine()`` in asyncpg event broker to only copy the
   connection options instead of directly using the engine
 - Simplified the MQTT event broker by providing a default ``client`` instance if omitted
+- Fixed ``CancelledError`` being reported as a crash on Python 3.7
 
 **4.0.0a1**